Partial Dependence Function

Estimates the partial dependence function of feature(s) v over a grid of values. Both multivariate and multivariable situations are supported.

partial_dep(object, ...)

# S3 method for default
partial_dep(
  object,
  v,
  X,
  pred_fun = stats::predict,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)

# S3 method for ranger
partial_dep(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)

# S3 method for Learner
partial_dep(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...), for instance type = "response" in a glm() model.

v

Vector of feature names.

X

A data.frame or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

grid

Optional evaluation grid in line with v.

grid_size

Controls the approximate grid size. If x has p columns, then each (non-discrete) column will be reduced to about the p-th root of grid_size values.

trim

A non-discrete numeric z is trimmed at these quantile probabilities before calculations. Set to c(0, 1) for no trimming.

strategy

How to find evaluation points of non-discrete numeric columns? Either "quantile" or "uniform" (via pretty()), see description of univariate_grid().

n_max

If X has more than n_max rows, a random sample of n_max rows is selected from X. In this case, set a random seed for reproducibility.

w

Optional vector of case weights for each row of X.

Value

A list with additional class "pd" containing the following elements

pd: data.frame representing both the evaluation grid and the corresponding partial dependencies.

v: Same as input v.

pred_names: Vector of column names representing partial dependencies.

Methods (by class)

partial_dep(default): Default method.

partial_dep(ranger): Method for "ranger" models.

partial_dep(Learner): Method for "mlr3" models.

Partial Dependence Functions

Let \(F: R^p \to R\) denote the prediction function that maps the \(p\)-dimensional feature vector \(\mathbf{x} = (x_1, \dots, x_p)\) to its prediction. Furthermore, let $$ F_s(\mathbf{x}_s) = E_{\mathbf{x}_{\setminus s}}(F(\mathbf{x}_s, \mathbf{x}_{\setminus s})) $$ be the partial dependence function of \(F\) on the feature subset \(\mathbf{x}_s\), where \(s \subseteq \{1, \dots, p\}\), as introduced in Friedman (2001). Here, the expectation runs over the joint marginal distribution of features \(\mathbf{x}_{\setminus s}\) not in \(\mathbf{x}_s\).

Given data, \(F_s(\mathbf{x}_s)\) can be estimated by the empirical partial dependence function

$$ \hat F_s(\mathbf{x}_s) = \frac{1}{n} \sum_{i = 1}^n F(\mathbf{x}_s, \mathbf{x}_{i\setminus s}), $$ where \(\mathbf{x}_{i\setminus s}\) \(i = 1, \dots, n\), are the observed values of \(\mathbf{x}_{\setminus s}\).

References

Friedman, Jerome H. "Greedy Function Approximation: A Gradient Boosting Machine." Annals of Statistics 29, no. 5 (2001): 1189-1232.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
pd <- partial_dep(fit, v = "Species", X = iris)
pd
#> First partial dependence rows (use x$pd to extract all)
#> 
#>      Species        y
#> 1     setosa 6.425687
#> 2 versicolor 5.702125
#> 3  virginica 5.402189

# Multivariable input
pd <- partial_dep(fit, v = c("Species", "Petal.Width"), X = iris)
pd
#> First partial dependence rows (use x$pd to extract all)
#> 
#>      Species Petal.Width        y
#> 1     setosa         0.1 6.772147
#> 2 versicolor         0.1 6.048585
#> 3  virginica         0.1 5.748649
#> 4     setosa         0.2 6.740632
#> 5 versicolor         0.2 6.017070
#> 6  virginica         0.2 5.717134

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
pd <- partial_dep(fit, v = "Petal.Width", X = iris)
pd
#> First partial dependence rows (use x$pd to extract all)
#> 
#>   Petal.Width Sepal.Length Sepal.Width
#> 1         0.1     5.849924    2.371959
#> 2         0.2     5.849325    2.434303
#> 3         0.3     5.848725    2.496648
#> 4         0.4     5.848126    2.558992
#> 5         0.5     5.847526    2.621337
#> 6         0.6     5.846927    2.683682
 
# MODEL THREE: Gamma GLM -> pass options to predict() via ...
fit <- glm(
  Sepal.Length ~ . + Petal.Width:Species, 
  data = iris, 
  family = Gamma(link = log)
)
partial_dep(fit, v = "Species", X = iris, type = "response")$pd
#>      Species        y
#> 1     setosa 6.989680
#> 2 versicolor 5.828742
#> 3  virginica 5.494072