Univariate Grid

Finds evaluation grid for any numeric or non-numeric vector z.

For discrete z (non-numeric, or numeric with at most grid_size unique values), this is simply sort(unique(z)).

Otherwise, if strategy = "quantile" (default), the evaluation points are computed as quantiles over a regular grid of probabilities from trim[1] to trim[2]. If strategy = "uniform", the evaluation points are the result of pretty() over the trimmed range of z. Set trim = c(0, 1) for no trimming.

Quantiles are calculated based on the inverse of the ECDF, i.e., with stats::quantile(..., type = 1).

univariate_grid(
  z,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform")
)

Arguments

z

A vector/factor.

grid_size

Approximate grid size.

trim

A non-discrete numeric z is trimmed at these quantile probabilities before calculations. Set to c(0, 1) for no trimming.

strategy

How to find evaluation points of non-discrete numeric columns? Either "quantile" or "uniform" (via pretty()), see description of univariate_grid().

Value

A vector/factor of evaluation points.

Examples

univariate_grid(iris$Species)
#> [1] setosa     versicolor virginica 
#> Levels: setosa versicolor virginica
univariate_grid(rev(iris$Species))                       # Same
#> [1] setosa     versicolor virginica 
#> Levels: setosa versicolor virginica

x <- iris$Sepal.Width
univariate_grid(x, grid_size = 5)                        # Quantile binning
#> [1] 2.2 2.8 3.0 3.3 4.2
univariate_grid(x, grid_size = 3, strategy = "uniform")  # Uniform pretty
#> [1] 2.0 2.5 3.0 3.5 4.0 4.5